Fixed bug: If the directory doesn't exist then don't bother moving it
git-svn-id: http://paistram/svn/repository/pdfcat/trunk@679 d73ca4c0-4f8e-dc11-baea-00112f300e52

+++ app/models/category.rb
@@ -23,7 +23,7 @@ class Category < ActiveRecord::Base
 				@newdir = STORE_DIR + "/" +  @clientname + "/" + name.downcase
 
 				# Move the category directory to the new one.
-				File.rename(@olddir,@newdir)
+				File.rename(@olddir,@newdir) if File.exists?(@olddir)
 			end
 		end
     end
